backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
pydantic>=2.0.0
python-multipart
# Database
tortoise-orm>=0.20.0
aerich>=0.7.0
# Agent
browser-use
langchain-openai
playwright>=1.40.0
langchain>=0.1.0
openai>=1.0.0
# Dev
pytest>=7.0.0
ruff>=0.1.0
